Modify: - 중복된  날짜데이터 가공 코드 컴포넌트 함수로 분리하여, import 해서 사용. 코드 수정. - 불필요한 주석 삭제
