+++ src/pages/Blog/components/BlogMain/BlogScrollArticle.tsx
@@ -3,6 +3,7 @@ import { useRecoilValue } from 'recoil';
 import { toggleSelector } from '../../../../recoil/ToggleState';
 import useAxios from '../../../../hooks/useAxios';
 import { BlogArticle } from './BlogArticle';
+import { blogCreatedAt } from '../../../../components/CreatedAt/CreatedAt';
 
 // Intersection Observer API 를 활용한 무한스크롤 페이지네이션
 export default function BlogScrollArticle() {
@@ -54,31 +55,12 @@ export default function BlogScrollArticle() {
     <>
       <section>
         {articles?.map((el: any) => {
-          console.log(articles);
           const nonSpoData = el.data.nonSpoPostData;
           const spoData = el?.data?.spoPostData;
           const spoToggleData = toggle ? spoData : nonSpoData;
           return spoToggleData?.map((el: any) => {
             const nickname = el.user.nickname;
-            // ** created_at UTC 날짜 가공하여, 며칠 전인지 구하는 로직
-            // 1. created_at 날짜를 new Date 객체로 변환 (한국표준시로 자동변경됨)
-            // 2. 지금 날짜와 created_at의 날짜의 오차를 구해서, date만 구해야 함
-            const dateString = el.created_at;
-            const utcToKst = new Date(dateString);
-            const createdY = utcToKst.getFullYear();
-            const createdM = utcToKst.getMonth();
-            const createdD = utcToKst.getDate();
-
-            const now = new Date();
-            const year = now.getFullYear();
-            const month = now.getMonth();
-            const date = now.getDate();
-
-            const createdDate = new Date(createdY, createdM, createdD);
-            const nowDate = new Date(year, month, date);
-
-            const cd_nd = nowDate.getTime() - createdDate.getTime(); // 오차구하기
-            const blogDate = cd_nd / (1000 * 60 * 60 * 24); // 밀리초로 나눔
+            const blogDate = blogCreatedAt(el.created_at);
 
             const {
               id,
